+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Add `CompletedIdentificationReport.name`
 - Add `ReportEndpoint.completedIdentifications` endpoint and `CompletedIdentificationReport` resource
 - Add `IdentificationResult` claims
+- Add `IdentificationsEndpoint.updateResult`
 
 ## [2.3.0] - 2024-03-28
 ### Added